🪜 : – Rename upper stairwell landing files
What / why:
- Align stairwell landing source and test filenames with the exported API.
- Remove stale upperLandingStub import paths without geometry changes.

How to test:
- npm run lint
- npm run typecheck
- npm run test:ci -- src/tests/movement/stairLayout.test.ts src/scene/structures/__tests__/floorTiles.test.ts src/scene/structures/__tests__/upperStairwellLanding.test.ts
- npm run test:ci
- npm run build
- npm run test:e2e -- playwright/small-screen-hud.spec.ts
