+++ src/scene/structures/upperStairwellLanding.ts
